:pencil2: fix: amount의 type을 전부 Long으로 수정
Former-commit-id: 5555d2abb1b4939aa01f03a6edb333be9ca57bff

+++ backend/src/main/java/com/sss/tally/api/calculate/dto/CalculateDto.java
@@ -150,9 +150,9 @@ public class CalculateDto {
 
 		private String paymentName;
 
-		private int myAmount;
+		private Long myAmount;
 
-		private int allAmount;
+		private Long allAmount;
 
 		private LocalDateTime paymentDate;
